[Update] News

+++ app/Http/Controllers/Admin/NewsController.php
@@ -39,18 +39,22 @@ class NewsController extends Controller
     }
 
     public function addNewsCate(Request $request){
-        $title      = "Thêm mới thể loại tin";
-        $errors     = NULL;
-        $listStatus = $this->getOption('listStatus');
-        $listNewsCate   = NewsCategory::getList();
-        $params     = $request->all();
+        $title        = "Thêm mới thể loại tin";
+        $errors       = NULL;
+        $listStatus   = $this->getOption('listStatus');
+        $listNewsCate = NewsCategory::getList();
+        $params       = $request->all();
         if ($request->isMethod('post')) {
-            $newNewsCate = NewsCategory::addNewsCate($params);
-            if($newNewsCate) {
-                $params  = null;
-                $errors = "Thêm thành công";
-            }else{
-                $errors = "Thêm thất bại";
+            if(empty($params['title'])) $errors['title'] = "Vui lòng nhập tên loại tin";
+            if(empty($params['seo_link'])) $errors['seo_link'] = "Vui lòng nhập seo link";
+            if(empty($errors)) {
+                $newNewsCate = NewsCategory::addNewsCate($params);
+                if($newNewsCate) {
+                    $params  = null;
+                    $errors['finish'] = "Thêm thành công";
+                }else{
+                    $errors['finish'] = "Thêm thất bại";
+                }
             }
         }
         return view("admin.news.addNewsCate")
@@ -64,13 +68,18 @@ class NewsController extends Controller
         $title  = "Sửa thể loại tin";
         $id     = (int) $request->route('id');
         $errors = NULL;
+        $newsCate = NewsCategory::findOrFail((int)$id)->toArray();
         $params = $request->all();
         if ($request->isMethod('post')) {
-            $errors = NewsCategory::editNewsCate($params);
-            if($errors) $errors = "Sủa thành công";
-            else        $errors = "Sửa thất bại";
+            if(empty($params['title'])) $errors['title'] = "Vui lòng nhập tên loại tin";
+            if(empty($params['seo_link'])) $errors['seo_link'] = "Vui lòng nhập seo link";
+            if(empty($errors)) {
+                $editNewsCate = NewsCategory::editNewsCate($params);
+                if($editNewsCate) $errors['finish'] = "Sủa thành công";
+                else $errors['finish'] = "Sửa thất bại";
+            }
+            $newsCate = $params;
         }
-        $newsCate = NewsCategory::findOrFail((int)$id);
         return view("admin.news.editNewsCate")
             ->with("title", $title)
             ->with("newsCate", $newsCate)
@@ -78,12 +87,26 @@ class NewsController extends Controller
     }
 
     public function delNewsCate($id){
-        NewsCategory::find((int)$id)->delete();
-        $back_url = redirect()->getUrlGenerator()->previous();
+        $existNews = News::where("id_cate","=" , $id)->count();
+        if(!$existNews) NewsCategory::find((int)$id)->delete();
+        $back_url  = redirect()->getUrlGenerator()->previous();
         return redirect()->guest($back_url);
     }
 
     // list news
+    public function status($id){
+        $news = News::findOrFail((int)$id);
+        if(!empty($news)){
+            if($news->status)
+                $news->status = 0;
+            else
+                $news->status = 1;
+        }
+        echo $news->save();
+        $back_url = redirect()->getUrlGenerator()->previous();
+        return redirect()->guest($back_url);
+    }
+
     public function index(Request $request){
         $title        = "Danh Sách Tin Tức";
         $page         = isset($request->page) ? $request->page : 1;
@@ -116,15 +139,19 @@ class NewsController extends Controller
         $listNewsCate = NewsCategory::getList();
         $params       = $request->all();
         if ($request->isMethod('post')) {
-            $clsImg = new Img();
-            $result = $clsImg->uploadImages();
-            foreach ($_FILES as $key => $value) $params[$key] = $result[$key];
-            $sNews = News::addNews($params);
-            if($sNews) {
-                $params = array();
-                $errors = "Thêm thành công";
-            }else{
-                $errors = "Thêm thất bại";
+            if(empty($params['title'])) $errors['title'] = "Vui lòng nhập tên loại tin";
+            if(empty($params['seo_link'])) $errors['seo_link'] = "Vui lòng nhập seo link";
+            if(empty($errors)) {
+                $clsImg = new Img();
+                $result = $clsImg->uploadImages();
+                foreach ($_FILES as $key => $value) $params[$key] = $result[$key];
+                $sNews = News::addNews($params);
+                if($sNews) {
+                    $params = array();
+                    $errors['finish'] = "Thêm thành công";
+                }else{
+                    $errors['finish'] = "Thêm thất bại";
+                }
             }
         }
         return view("admin.news.add")
@@ -139,42 +166,81 @@ class NewsController extends Controller
         $id           = (int) $request->route('id');
         $cls_img      = new Img();
         $errors       = NULL;
-        $listColors   = Colors::getList();
-        $listCategory = ProductCategory::getList();
-        $listType     = ProductType::getList();
-        $getProduct   = Product::findOrFail((int)$id);
+        $listNewsCate = NewsCategory::getList();
+        $getNews      = News::findOrFail((int)$id)->toArray();
         $params       = $request->all();
         if ($request->isMethod('post')) {
-            $result = $cls_img->uploadImages();
-            foreach ($_FILES as $key => $value) {
-                if($params[$key."_url"] != $result[$key] && !empty($result[$key])){
-                    $params[$key] = $result[$key];
-                    $cls_img->removeImages($getProduct[$key]);
+            if(empty($params['title'])) $errors['title'] = "Vui lòng nhập tên loại tin";
+            if(empty($params['seo_link'])) $errors['seo_link'] = "Vui lòng nhập seo link";
+            if(empty($errors)) {
+                $result = $cls_img->uploadImages();
+                foreach ($_FILES as $key => $value) {
+                    if($params[$key."_url"] != $result[$key] && !empty($result[$key])){
+                        $params[$key] = $result[$key];
+                        $cls_img->removeImages($getNews[$key]);
+                    }
+                    else {
+                        if(empty($params[$key."_url"])) $cls_img->removeImages($getNews[$key]);
+                        $params[$key] = $params[$key."_url"];
+                    }
                 }
-                else {
-                    if(empty($params[$key."_url"])) $cls_img->removeImages($getProduct[$key]);
-                    $params[$key] = $params[$key."_url"];
-                }
+                $editNews = News::editNews($params);
+                if($editNews) $errors['finish'] = "Sủa thành công";
+                else $errors['finish'] = "Sửa thất bại";
             }
-            $errors = Product::editProduct($params);
-            if($errors) $errors = "Sủa thành công";
-            else        $errors = "Sửa thất bại";
+            $getNews = $params;
          }
-        // reload data again
-        $getProduct   = Product::findOrFail((int)$id);
-        return view("admin.product.editProduct")
+        return view("admin.news.edit")
             ->with("title" , $title )
-            ->with("product", $getProduct)
-            ->with("listColors", $listColors)
-            ->with("listCategory", $listCategory)
-            ->with("listType", $listType)
+            ->with("news", $getNews)
+            ->with("listNewsCate", $listNewsCate)
             ->with("errors", $errors);
     }
 
-    public function delProduct($id){
-        Product::find((int)$id)->delete();
+    public function del($id){
+        News::find((int)$id)->delete();
         $back_url = redirect()->getUrlGenerator()->previous();
         return redirect()->guest($back_url);
+    }
+
+    public function sort(Request $request){
+        $title        = "Sắp Xếp Tin Tức";
+        $page         = isset($request->page) ? $request->page : 1;
+        $builder      = News::select();
+        $listNewsCate = NewsCategory::getList();
+        $builder->orderByRaw("ord ASC, id DESC");
+        $totalResult  = $builder->count();
+        $listNews  = $builder->paginate(LIMIT_PAGE);
+
+        return view("admin.news.sort")
+            ->with("title", $title)
+            ->with("page", $page)
+            ->with("listNewsCate", $listCategory)
+            ->with("listNews", $listNews);
+    }
+
+    public function ajaxSort(Request $request){
+        $params = $request->all();
+        if(!empty($params)) {
+            parse_str($params['data'],$data);
+            $page      = (int) $params['page'];
+            $limitPage = LIMIT_PAGE;
+            $offset    = $limitPage * ($page - 1);
+            $listNews = News::orderByRaw("ord ASC, id DESC")->get();
+            foreach ($listNews as $key => $value) {
+                $news = News::find($value->id);
+                $news->ord = $key+1;
+                $news->save();
+            }
+
+            foreach ($data['item'] as $key => $value) {
+                $key += $offset;
+                $news = News::find($value);
+                $news->ord = $key;
+                $news->save();
+            }
+        }
+        return response()->json(array("status" => true));
     }
